refactor: change server folder structure
